+++ src/main/scala/com/kodekutters/czml/czmlProperties.scala
@@ -30,8 +30,6 @@
 
 package com.kodekutters.czml
 
-import com.typesafe.scalalogging.Logger
-import org.slf4j.LoggerFactory
 import play.api.libs.functional.syntax._
 import play.api.libs.json.Reads._
 import play.api.libs.json._
@@ -59,8 +57,6 @@ import scala.collection.immutable.ListMap
   * the CZMLPacket object and all its constituent czml properties.
   */
 package object czmlProperties {
-
-  val logger = Logger(LoggerFactory.getLogger("czmlProperties"))
 
   /**
     * all CzmlPacket constituent properties extend this trait
@@ -1086,9 +1082,7 @@ package object czmlProperties {
     def apply[T <: Packet](doc: String)(implicit fmt: Reads[T]): CZML[T] = {
       Json.parse(doc).asOpt[JsArray] match {
         case Some(jsonArray) => CZML(jsonArray)
-        case None =>
-          logger.error("could not parse the CZML document")
-          new CZML[T](ArrayBuffer[T]())
+        case None => new CZML[T](ArrayBuffer[T]())
       }
     }
 
@@ -1106,9 +1100,7 @@ package object czmlProperties {
             val listBuf = ArrayBuffer.empty ++= (for (p <- list) yield Json.fromJson(p)(fmt).asOpt)
             JsSuccess(new CZML[T](listBuf.flatten))
 
-          case _ =>
-            logger.error("could not read the CZML document")
-            JsSuccess(new CZML[T](ArrayBuffer[T]()))
+          case _ => JsSuccess(new CZML[T](ArrayBuffer[T]()))
         }
       }
     }